@@ -4,28 +4,101 @@ const fs = require('fs');
 const path = require('path');
 
 // From https://threejs.org/examples/files.js
+// or generated from https://github.com/mrdoob/three.js/tree/dev/examples
 const names = [
+	'css2d_label',
+	'css3d_molecules',
+	'css3d_orthographic',
+	'css3d_panorama',
+	'css3d_panorama_deviceorientation',
+	'css3d_periodictable',
+	'css3d_sandbox',
+	'css3d_sprites',
+	'css3d_youtube',
+	'index',
+	'misc_animation_authoring',
+	'misc_animation_groups',
+	'misc_animation_keys',
+	'misc_boxselection',
+	'misc_controls_deviceorientation',
+	'misc_controls_fly',
+	'misc_controls_map',
+	'misc_controls_orbit',
+	'misc_controls_pointerlock',
+	'misc_controls_trackball',
+	'misc_controls_transform',
+	'misc_exporter_collada',
+	'misc_exporter_draco',
+	'misc_exporter_gltf',
+	'misc_exporter_obj',
+	'misc_exporter_stl',
+	'misc_fps',
+	'misc_lookat',
+	'misc_uv_tests',
+	'raytracing_sandbox',
+	'software_geometry_earth',
+	'software_lines_splines',
+	'software_sandbox',
+	'svg_lines',
+	'svg_sandbox',
+	'webaudio_orientation',
+	'webaudio_sandbox',
+	'webaudio_timing',
+	'webaudio_visualizer',
+	'webgl2_materials_texture2darray',
+	'webgl2_materials_texture3d',
+	'webgl2_multisampled_renderbuffers',
+	'webgl2_sandbox',
+	'webgldeferred_animation',
 	'webgl_animation_cloth',
-	'webgl_animation_scene',
+	'webgl_animation_keyframes',
+	'webgl_animation_multiple',
 	'webgl_animation_skinning_blending',
 	'webgl_animation_skinning_morph',
+	'webgl_buffergeometry',
+	'webgl_buffergeometry_constructed_from_geometry',
+	'webgl_buffergeometry_custom_attributes_particles',
+	'webgl_buffergeometry_drawcalls',
+	'webgl_buffergeometry_indexed',
+	'webgl_buffergeometry_instancing',
+	'webgl_buffergeometry_instancing2',
+	'webgl_buffergeometry_instancing_billboards',
+	'webgl_buffergeometry_instancing_dynamic',
+	'webgl_buffergeometry_instancing_interleaved_dynamic',
+	'webgl_buffergeometry_instancing_lambert',
+	'webgl_buffergeometry_lines',
+	'webgl_buffergeometry_lines_indexed',
+	'webgl_buffergeometry_morphtargets',
+	'webgl_buffergeometry_points',
+	'webgl_buffergeometry_points_interleaved',
+	'webgl_buffergeometry_rawshader',
+	'webgl_buffergeometry_selective_draw',
+	'webgl_buffergeometry_uint',
 	'webgl_camera',
+	'webgl_camera_array',
 	'webgl_camera_cinematic',
 	'webgl_camera_logarithmicdepthbuffer',
 	'webgl_clipping',
 	'webgl_clipping_advanced',
 	'webgl_clipping_intersection',
+	'webgl_custom_attributes',
+	'webgl_custom_attributes_lines',
+	'webgl_custom_attributes_points',
+	'webgl_custom_attributes_points2',
+	'webgl_custom_attributes_points3',
 	'webgl_decals',
 	'webgl_depth_texture',
 	'webgl_effects_anaglyph',
+	'webgl_effects_ascii',
 	'webgl_effects_parallaxbarrier',
 	'webgl_effects_peppersghost',
 	'webgl_effects_stereo',
-	'webgl_exporter_obj',
+	'webgl_fire',
+	'webgl_framebuffer_texture',
+	'webgl_furnace_test',
 	'webgl_geometries',
-	'webgl_geometries2',
+	'webgl_geometries_parametric',
 	'webgl_geometry_colors',
-	'webgl_geometry_colors_blender',
 	'webgl_geometry_colors_lookuptable',
 	'webgl_geometry_convex',
 	'webgl_geometry_cube',
@@ -46,40 +119,43 @@ const names = [
 	'webgl_geometry_terrain_fog',
 	'webgl_geometry_terrain_raycast',
 	'webgl_geometry_text',
-	'webgl_geometry_text_earcut',
-	'webgl_geometry_text_pnltri',
+	'webgl_geometry_text_shapes',
+	'webgl_geometry_text_stroke',
 	'webgl_gpgpu_birds',
-	'webgl_gpgpu_water',
 	'webgl_gpgpu_protoplanet',
-	'webgl_gpu_particle_system',
-	'webgl_hdr',
+	'webgl_gpgpu_water',
 	'webgl_helpers',
 	'webgl_interactive_buffergeometry',
 	'webgl_interactive_cubes',
 	'webgl_interactive_cubes_gpu',
-	'webgl_interactive_instances_gpu',
 	'webgl_interactive_cubes_ortho',
 	'webgl_interactive_draggablecubes',
+	'webgl_interactive_instances_gpu',
 	'webgl_interactive_lines',
 	'webgl_interactive_points',
 	'webgl_interactive_raycasting_points',
 	'webgl_interactive_voxelpainter',
 	'webgl_kinect',
+	'webgl_layers',
 	'webgl_lensflares',
+	'webgl_lightningstrike',
+	'webgl_lightprobe',
+	'webgl_lightshafts',
 	'webgl_lights_hemisphere',
 	'webgl_lights_physical',
 	'webgl_lights_pointlights',
 	'webgl_lights_pointlights2',
+	'webgl_lights_rectarealight',
 	'webgl_lights_spotlight',
 	'webgl_lights_spotlights',
-	'webgl_lights_arealight',
-	'webgl_lights_rectarealight',
 	'webgl_lines_colors',
-	'webgl_lines_cubes',
 	'webgl_lines_dashed',
+	'webgl_lines_fat',
+	'webgl_lines_fat_wireframe',
 	'webgl_lines_sphere',
-	'webgl_lines_splines',
+	'webgl_loader_3ds',
 	'webgl_loader_3mf',
+	'webgl_loader_3mf_materials',
 	'webgl_loader_amf',
 	'webgl_loader_assimp',
 	'webgl_loader_assimp2json',
@@ -87,41 +163,63 @@ const names = [
 	'webgl_loader_babylon',
 	'webgl_loader_bvh',
 	'webgl_loader_collada',
-	'webgl_loader_collada_keyframe',
 	'webgl_loader_collada_kinematics',
 	'webgl_loader_collada_skinning',
 	'webgl_loader_ctm',
 	'webgl_loader_ctm_materials',
+	'webgl_loader_draco',
 	'webgl_loader_fbx',
+	'webgl_loader_fbx_nurbs',
+	'webgl_loader_gcode',
 	'webgl_loader_gltf',
-	'webgl_loader_json_blender',
+	'webgl_loader_gltf_extensions',
+	'webgl_loader_imagebitmap',
 	'webgl_loader_json_claraio',
-	'webgl_loader_json_objconverter',
+	'webgl_loader_kmz',
+	'webgl_loader_ldraw',
+	'webgl_loader_lwo',
 	'webgl_loader_md2',
 	'webgl_loader_md2_control',
 	'webgl_loader_mmd',
-	'webgl_loader_mmd_pose',
 	'webgl_loader_mmd_audio',
-	'webgl_loader_msgpack',
-	'webgl_loader_obj',
-	'webgl_loader_obj_mtl',
+	'webgl_loader_mmd_pose',
+	'webgl_loader_nodes',
 	'webgl_loader_nrrd',
+	'webgl_loader_obj',
+	'webgl_loader_obj2',
+	'webgl_loader_obj2_meshspray',
+	'webgl_loader_obj2_options',
+	'webgl_loader_obj2_run_director',
+	'webgl_loader_obj_mtl',
 	'webgl_loader_pcd',
 	'webgl_loader_pdb',
+	'webgl_loader_playcanvas',
 	'webgl_loader_ply',
-	'webgl_loader_ttf',
+	'webgl_loader_prwm',
 	'webgl_loader_sea3d',
+	'webgl_loader_sea3d_bvh',
+	'webgl_loader_sea3d_bvh_retarget',
 	'webgl_loader_sea3d_hierarchy',
 	'webgl_loader_sea3d_keyframe',
 	'webgl_loader_sea3d_morph',
 	'webgl_loader_sea3d_physics',
 	'webgl_loader_sea3d_skinning',
 	'webgl_loader_sea3d_sound',
-	'webgl_loader_scene',
 	'webgl_loader_stl',
-	'webgl_loader_utf8',
+	'webgl_loader_svg',
+	'webgl_loader_texture_basis',
+	'webgl_loader_texture_dds',
+	'webgl_loader_texture_exr',
+	'webgl_loader_texture_hdr',
+	'webgl_loader_texture_ktx',
+	'webgl_loader_texture_pvrtc',
+	'webgl_loader_texture_rgbm',
+	'webgl_loader_texture_tga',
+	'webgl_loader_ttf',
+	'webgl_loader_vrm',
 	'webgl_loader_vrml',
 	'webgl_loader_vtk',
+	'webgl_loader_x',
 	'webgl_lod',
 	'webgl_marchingcubes',
 	'webgl_materials',
@@ -131,213 +229,156 @@ const names = [
 	'webgl_materials_bumpmap_skin',
 	'webgl_materials_cars',
 	'webgl_materials_channels',
+	'webgl_materials_compile',
 	'webgl_materials_cubemap',
 	'webgl_materials_cubemap_balls_reflection',
 	'webgl_materials_cubemap_balls_refraction',
 	'webgl_materials_cubemap_dynamic',
-	'webgl_materials_cubemap_dynamic2',
 	'webgl_materials_cubemap_refraction',
+	'webgl_materials_curvature',
 	'webgl_materials_displacementmap',
 	'webgl_materials_envmaps',
+	'webgl_materials_envmaps_exr',
 	'webgl_materials_envmaps_hdr',
+	'webgl_materials_envmaps_hdr_nodes',
+	'webgl_materials_envmaps_parallax',
 	'webgl_materials_grass',
 	'webgl_materials_lightmap',
+	'webgl_materials_matcap',
+	'webgl_materials_modified',
 	'webgl_materials_nodes',
 	'webgl_materials_normalmap',
+	'webgl_materials_normalmap_object_space',
 	'webgl_materials_parallaxmap',
 	'webgl_materials_reflectivity',
 	'webgl_materials_shaders_fresnel',
 	'webgl_materials_skin',
 	'webgl_materials_standard',
 	'webgl_materials_texture_anisotropy',
-	'webgl_materials_texture_compressed',
+	'webgl_materials_texture_canvas',
 	'webgl_materials_texture_filters',
-	'webgl_materials_texture_hdr',
 	'webgl_materials_texture_manualmipmap',
-	'webgl_materials_texture_pvrtc',
-	'webgl_materials_texture_tga',
+	'webgl_materials_texture_partialupdate',
+	'webgl_materials_texture_rotation',
+	'webgl_materials_translucency',
 	'webgl_materials_transparency',
 	'webgl_materials_variations_basic',
 	'webgl_materials_variations_lambert',
 	'webgl_materials_variations_phong',
-	'webgl_materials_variations_standard',
 	'webgl_materials_variations_physical',
+	'webgl_materials_variations_standard',
 	'webgl_materials_variations_toon',
 	'webgl_materials_video',
+	'webgl_materials_video_webcam',
 	'webgl_materials_wireframe',
+	'webgl_math_orientation_transform',
 	'webgl_mirror',
 	'webgl_mirror_nodes',
 	'webgl_modifier_simplifier',
 	'webgl_modifier_subdivision',
 	'webgl_modifier_tessellation',
-	'webgl_morphnormals',
 	'webgl_morphtargets',
 	'webgl_morphtargets_horse',
-	'webgl_morphtargets_human',
+	'webgl_morphtargets_sphere',
 	'webgl_multiple_canvases_circle',
 	'webgl_multiple_canvases_complex',
 	'webgl_multiple_canvases_grid',
 	'webgl_multiple_elements',
 	'webgl_multiple_elements_text',
 	'webgl_multiple_renderers',
+	'webgl_multiple_scenes_comparison',
 	'webgl_multiple_views',
 	'webgl_nearestneighbour',
-	'webgl_objects_update',
-	'webgl_octree',
-	'webgl_octree_raycasting',
 	'webgl_panorama_cube',
 	'webgl_panorama_dualfisheye',
 	'webgl_panorama_equirectangular',
 	'webgl_performance',
 	'webgl_performance_doublesided',
+	'webgl_performance_nodes',
 	'webgl_performance_static',
-	'webgl_physics_terrain',
-	'webgl_physics_rope',
 	'webgl_physics_cloth',
-	'webgl_physics_volume',
 	'webgl_physics_convex_break',
+	'webgl_physics_rope',
+	'webgl_physics_terrain',
+	'webgl_physics_volume',
 	'webgl_points_billboards',
-	'webgl_points_billboards_colors',
 	'webgl_points_dynamic',
-	'webgl_points_random',
 	'webgl_points_sprites',
+	'webgl_points_waves',
 	'webgl_postprocessing',
 	'webgl_postprocessing_advanced',
+	'webgl_postprocessing_afterimage',
 	'webgl_postprocessing_backgrounds',
 	'webgl_postprocessing_crossfade',
 	'webgl_postprocessing_dof',
 	'webgl_postprocessing_dof2',
+	'webgl_postprocessing_fxaa',
 	'webgl_postprocessing_glitch',
 	'webgl_postprocessing_godrays',
 	'webgl_postprocessing_masking',
+	'webgl_postprocessing_nodes',
+	'webgl_postprocessing_nodes_pass',
+	'webgl_postprocessing_outline',
+	'webgl_postprocessing_pixel',
+	'webgl_postprocessing_procedural',
+	'webgl_postprocessing_rgb_halftone',
+	'webgl_postprocessing_sao',
+	'webgl_postprocessing_smaa',
+	'webgl_postprocessing_sobel',
 	'webgl_postprocessing_ssaa',
 	'webgl_postprocessing_ssaa_unbiased',
-	'webgl_postprocessing_nodes',
-	'webgl_postprocessing_outline',
-	'webgl_postprocessing_procedural',
-	'webgl_postprocessing_smaa',
 	'webgl_postprocessing_ssao',
 	'webgl_postprocessing_taa',
 	'webgl_postprocessing_unreal_bloom',
+	'webgl_postprocessing_unreal_bloom_selective',
+	'webgl_raycast_sprite',
 	'webgl_raycast_texture',
+	'webgl_raymarching_reflect',
 	'webgl_read_float_buffer',
+	'webgl_refraction',
 	'webgl_rtt',
 	'webgl_sandbox',
 	'webgl_shader',
-	'webgl_shader_lava',
 	'webgl_shader2',
 	'webgl_shaders_ocean',
 	'webgl_shaders_ocean2',
 	'webgl_shaders_sky',
 	'webgl_shaders_tonemapping',
 	'webgl_shaders_vector',
+	'webgl_shader_lava',
 	'webgl_shading_physical',
 	'webgl_shadowmap',
+	'webgl_shadowmap_pcss',
 	'webgl_shadowmap_performance',
 	'webgl_shadowmap_pointlight',
 	'webgl_shadowmap_viewer',
 	'webgl_shadowmesh',
+	'webgl_simple_gi',
 	'webgl_skinning_simple',
 	'webgl_sprites',
+	'webgl_sprites_nodes',
 	'webgl_terrain_dynamic',
 	'webgl_test_memory',
 	'webgl_test_memory2',
+	'webgl_tiled_forward',
 	'webgl_tonemapping',
 	'webgl_trails',
 	'webgl_video_panorama_equirectangular',
-	'webgl_buffergeometry',
-	'webgl_buffergeometry_constructed_from_geometry',
-	'webgl_buffergeometry_custom_attributes_particles',
-	'webgl_buffergeometry_drawcalls',
-	'webgl_buffergeometry_instancing',
-	'webgl_buffergeometry_instancing_billboards',
-	'webgl_buffergeometry_instancing_dynamic',
-	'webgl_buffergeometry_instancing_interleaved_dynamic',
-	'webgl_buffergeometry_lines',
-	'webgl_buffergeometry_lines_indexed',
-	'webgl_buffergeometry_points',
-	'webgl_buffergeometry_points_interleaved',
-	'webgl_buffergeometry_rawshader',
-	'webgl_buffergeometry_selective_draw',
-	'webgl_buffergeometry_uint',
-	'webgl_custom_attributes',
-	'webgl_custom_attributes_lines',
-	'webgl_custom_attributes_points',
-	'webgl_custom_attributes_points2',
-	'webgl_custom_attributes_points3',
-	'webgl_raymarching_reflect',
-	'webgldeferred_animation',
-	'webgl2_sandbox',
+	'webgl_water',
+	'webgl_water_flowmap',
+	'webgl_worker_offscreencanvas',
+	'webvr_ballshooter',
 	'webvr_cubes',
+	'webvr_dragging',
+	'webvr_lorenzattractor',
+	'webvr_paint',
 	'webvr_panorama',
 	'webvr_rollercoaster',
-	'webvr_shadow',
+	'webvr_sandbox',
+	'webvr_sculpt',
 	'webvr_video',
-	'webvr_vive',
-	'webvr_vive_dragging',
 	'webvr_vive_paint',
-	'webvr_vive_sculpt',
-	'css3d_molecules',
-	'css3d_panorama',
-	'css3d_panorama_deviceorientation',
-	'css3d_periodictable',
-	'css3d_sandbox',
-	'css3d_sprites',
-	'css3d_youtube',
-	'misc_animation_authoring',
-	'misc_animation_keys',
-	'misc_controls_deviceorientation',
-	'misc_controls_fly',
-	'misc_controls_orbit',
-	'misc_controls_pointerlock',
-	'misc_controls_trackball',
-	'misc_controls_transform',
-	'misc_fps',
-	'misc_lights_test',
-	'misc_lookat',
-	'misc_sound',
-	'misc_ubiquity_test',
-	'misc_ubiquity_test2',
-	'misc_uv_tests',
-	'canvas_ascii_effect',
-	'canvas_camera_orthographic',
-	'canvas_camera_orthographic2',
-	'canvas_geometry_birds',
-	'canvas_geometry_cube',
-	'canvas_geometry_earth',
-	'canvas_geometry_hierarchy',
-	'canvas_geometry_nurbs',
-	'canvas_geometry_panorama',
-	'canvas_geometry_panorama_fisheye',
-	'canvas_geometry_shapes',
-	'canvas_geometry_terrain',
-	'canvas_geometry_text',
-	'canvas_interactive_cubes',
-	'canvas_interactive_cubes_tween',
-	'canvas_interactive_particles',
-	'canvas_interactive_voxelpainter',
-	'canvas_lights_pointlights',
-	'canvas_lines',
-	'canvas_lines_colors',
-	'canvas_lines_colors_2d',
-	'canvas_lines_dashed',
-	'canvas_lines_sphere',
-	'canvas_materials',
-	'canvas_materials_normal',
-	'canvas_materials_reflection',
-	'canvas_materials_video',
-	'canvas_morphtargets_horse',
-	'canvas_particles_floor',
-	'canvas_particles_random',
-	'canvas_particles_sprites',
-	'canvas_particles_waves',
-	'canvas_performance',
-	'canvas_sandbox',
-	'raytracing_sandbox',
-	'software_geometry_earth',
-	'software_lines_splines',
-	'software_sandbox',
-	'svg_sandbox'
+	'webvr_vive_sculpt'
 ];
 
 
@@ -351,7 +392,7 @@ describe('Three.js examples', () => {
 			browser.sleep(2500);
 			browser.takeScreenshot().then(data => {
 				const stream = fs.createWriteStream(path.join(__dirname, `screenshots/${id}.png`));
-				stream.write(new Buffer(data, 'base64'));
+				stream.write(Buffer.from(data, 'base64'));
 				stream.end();
 			});
 		});
